+++ CHANGES.md
@@ -44,3 +44,15 @@
 - **[2026-05-06 18:21:26]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\admin\index.php`
 - **[2026-05-06 18:21:38]** `Write` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\.htaccess`
 - **[2026-05-06 18:21:40]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\Dockerfile`
+- **[2026-05-06 18:58:23]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\includes\db.php`
+- **[2026-05-06 18:58:28]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\includes\db.php`
+- **[2026-05-06 18:58:45]** `Write` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\admin\settings.php`
+- **[2026-05-06 18:58:52]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\admin\auth.php`
+- **[2026-05-06 18:59:03]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\index.php`
+- **[2026-05-06 18:59:21]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\index.php`
+- **[2026-05-06 18:59:30]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\index.php`
+- **[2026-05-06 18:59:38]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\index.php`
+- **[2026-05-06 18:59:51]** `Write` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\favicon.svg`
+- **[2026-05-06 18:59:59]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\product.php`
+- **[2026-05-06 19:00:06]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\admin\index.php`
+- **[2026-05-06 19:00:14]** `Edit` → `C:\Users\renau\Documents\GitHub\LyblicsNoel\admin\edit.php`